added the final view from game but not dinamic
