+++ gameMath/Views/InitPopUpView.swift
